Allow hot-reloading of the plugin
When you save boot.py, all modified py files should be picked up now
